Add class for status bar menu
